evans: lazy-load (code-split) checkout
